Standardize naming of test cases
Standardized the naming of the test cases, both in the filename and in the test case class names.

For the filename the full type name should be appended with `_Tests`. For the test case class name any illegal characters are simply replaced with `_`.

Filename: TypeTo.Be+Tested_Tests.swift
Test class name: TypeTo_Be_Tested_Tests